+++ src/decimal-field.tsx
@@ -150,7 +150,7 @@ const DecimalField = ({
             {...fieldProps}
             inputComponent={DecimalInput}
             inputProps={{maxDecimals}}
-            dataAttributes={{'component-name': 'DecimalField', ...dataAttributes}}
+            dataAttributes={{'component-name': 'DecimalField', testid: 'DecimalField', ...dataAttributes}}
         />
     );
 };